[lib] Updated data type of fixed table entries to int...

+++ tehreer-android/src/main/java/com/mta/tehreer/opentype/FontHeaderTable.java
@@ -66,12 +66,12 @@ public class FontHeaderTable {
         this.table = new StructTable(typeface, pointer);
     }
 
-    public long version() {
-        return table.readUInt32(VERSION);
+    public int version() {
+        return table.readInt32(VERSION);
     }
 
-    public long fontRevision() {
-        return table.readUInt32(FONT_REVISION);
+    public int fontRevision() {
+        return table.readInt32(FONT_REVISION);
     }
 
     public long checkSumAdjustment() {
