+++ tehreer-android/src/main/java/com/mta/tehreer/opentype/HorizontalHeaderTable.java
@@ -62,8 +62,8 @@ public class HorizontalHeaderTable {
         this.table = new StructTable(typeface, pointer);
     }
 
-    public long version() {
-        return table.readUInt32(VERSION);
+    public int version() {
+        return table.readInt32(VERSION);
     }
 
     public short ascender() {
